+++ src/core/Command.ts
@@ -1,8 +1,14 @@
 export default abstract class Command {
-  public name!: string
-  public decsription!: string
+  public name: string
+  public decsription: string
   public aliases!: string[]
+  public cooldown!: number
   public guildOnly!: boolean
+  public ownerOnly!: boolean
+  public slashCommandOptions!: ISlashOption[]
+  public userPermissions!: string[]
+  public clientPermissions!: string[]
+  public usage!: string
 
   public abstract run(client, message, args): any;
